Default view & navigation #6

+++ server/src/main/java/com/bethibande/repository/web/api/RepositoryEndpoint.java
@@ -1,13 +1,23 @@
 package com.bethibande.repository.web.api;
 
-import com.bethibande.repository.jpa.repository.*;
+import com.bethibande.repository.jpa.artifact.Artifact;
+import com.bethibande.repository.jpa.artifact.ArtifactVersion;
+import com.bethibande.repository.jpa.repository.Repository;
+import com.bethibande.repository.jpa.repository.RepositoryDTO;
+import com.bethibande.repository.jpa.repository.RepositoryDTOWithoutId;
+import com.bethibande.repository.jpa.repository.RepositoryDTOWithoutSettings;
 import com.bethibande.repository.web.CRUDResponse;
+import jakarta.annotation.security.PermitAll;
+import jakarta.annotation.security.RolesAllowed;
 import jakarta.transaction.Transactional;
+import jakarta.validation.constraints.NotNull;
 import jakarta.ws.rs.*;
 
+import java.time.Instant;
 import java.util.List;
 
-@Path("/api/v1/")
+@RolesAllowed("ADMIN")
+@Path("/api/v1/repository")
 public class RepositoryEndpoint {
 
     @POST
@@ -15,13 +25,11 @@ public class RepositoryEndpoint {
     public RepositoryDTO create(final RepositoryDTOWithoutId dto) {
         final Repository repository = new Repository();
         repository.name = dto.name();
-        repository.type = dto.type();
+        repository.packageManager = dto.packageManager();
         repository.settings = dto.settings();
-        repository.backend = RepositoryBackend.findById(dto.backendId());
-
-        if (repository.backend == null) throw new BadRequestException("Backend id unknown");
 
         repository.persist();
+
         return RepositoryDTO.from(repository);
     }
 
@@ -29,31 +37,61 @@ public class RepositoryEndpoint {
     @Transactional
     public RepositoryDTO update(final RepositoryDTO dto) {
         final Repository repository = Repository.findById(dto.id());
-        if (repository == null) throw new NotFoundException();
+        if (repository == null) throw new NotFoundException("Unknown repository");
 
         repository.name = dto.name();
-        repository.type = dto.type();
+        repository.packageManager = dto.packageManager();
         repository.settings = dto.settings();
-        repository.backend = RepositoryBackend.findById(dto.backendId());
+        repository.persist();
 
-        if (repository.backend == null) throw new BadRequestException("Backend id unknown");
         return RepositoryDTO.from(repository);
     }
 
     @GET
     @Transactional
-    public List<RepositoryDTOWithoutBackend> findAll() {
+    public List<RepositoryDTO> list() {
         return Repository.<Repository>listAll()
                 .stream()
-                .map(RepositoryDTOWithoutBackend::from)
+                .map(RepositoryDTO::from)
+                .toList();
+    }
+
+    public record RepositoryOverviewDTO(
+            @NotNull RepositoryDTOWithoutSettings repository,
+            long artifactsCount,
+            Instant lastUpdated
+    ) {
+    }
+
+    @GET
+    @PermitAll
+    @Transactional
+    @Path("/overview")
+    public List<RepositoryOverviewDTO> overview() {
+
+        final List<RepositoryDTOWithoutSettings> repositories = Repository.<Repository>listAll()
+                .stream()
+                .map(RepositoryDTOWithoutSettings::from)
+                .toList();
+
+        return repositories.stream()
+                .map(repository -> new RepositoryOverviewDTO(
+                        repository,
+                        Artifact.count("repository.id = ?1", repository.id()),
+                        ArtifactVersion.findMaxUpdated(repository.id())
+                ))
                 .toList();
     }
 
     @DELETE
-    public CRUDResponse<Void> delete(final long id) {
+    @Transactional
+    @Path("/{id}")
+    public CRUDResponse<Void> delete(@PathParam("id") final Long id) {
+        if (Artifact.count("repository.id = ?", id) > 0) return CRUDResponse.failure("Cannot delete repository with artifacts", "repository.delete.artifacts");
+
         return Repository.deleteById(id)
                 ? CRUDResponse.success(null)
-                : CRUDResponse.failure("Not found", "error.notFound");
+                : CRUDResponse.failure("Unknown repository", "repository.delete.unknown");
     }
 
 }
